backend/api/rest/views.py
from api.rest.serializers import UserSerializer
from accounts.models import User
from rest_framework import viewsets


class UserViewSet(viewsets.ModelViewSet):
    queryset = User.objects.all()
    serializer_class = UserSerializer
