backend/requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
decorator==4.1.2
defusedxml==0.5.0
Django==1.11.7
django-allauth==0.34.0
django-autocomplete-light==3.2.10
django-extensions==1.9.1
django-filter==1.1.0
django-rest-auth==0.9.2
django-rest-framework==0.1.0
djangorestframework==3.6.4
graphene==2.0
graphene-django==2.0.0
graphql-core==2.0
graphql-relay==0.4.5
idna==2.6
ipdb==0.10.3
ipython==6.2.1
ipython-genutils==0.2.0
iso8601==0.1.12
jedi==0.11.0
Markdown==2.6.9
oauthlib==2.0.6
olefile==0.44
parso==0.1.0
pexpect==4.2.1
pickleshare==0.7.4
Pillow==4.3.0
promise==2.1
prompt-toolkit==1.0.15
psycopg2==2.7.3.1
ptyprocess==0.5.2
Pygments==2.2.0
python3-openid==3.1.0
pytz==2017.3
requests==2.18.4
requests-oauthlib==0.8.0
Rx==1.6.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.11.0
traitlets==4.3.2
typing==3.6.2
urllib3==1.22
wcwidth==0.1.7
